+++ contracts/interface/IStrategy.sol
@@ -8,9 +8,9 @@ interface IStrategy is IModule {
     
     function init() external;
     function execute() external;
-    function setAllowance(address token, uint amount) external;
-    function toVault(address token, uint amount) external;
-    function fromVault(address token, uint amount) external;
+    function setAllowance(address token, uint256 amount) external;
+    function toVault(address token, uint256 amount) external;
+    function fromVault(address token, uint256 amount) external;
     function closeAllPositions() external returns(bool);
     function closePositions(address token) external returns(bool success);
     function oneToken() external view returns(address);
