+++ contracts/oracle/composite/ICHICompositeOracle.sol
@@ -12,7 +12,7 @@ import "../../_openzeppelin/math/SafeMath.sol";
 
 contract ICHICompositeOracle is OracleCommon {
 
-    using SafeMath for uint;
+    using SafeMath for uint256;
     
     address[] public oracleContracts;
     address[] public interimTokens;
@@ -38,7 +38,7 @@ contract ICHICompositeOracle is OracleCommon {
      @dev there is nothing to do. Deploy separate instances configured for distinct baseTokens
      */
     function init(address baseToken) external onlyModuleOrFactory override {
-        for(uint i=0; i<oracleContracts.length; i++) {
+        for(uint256 i=0; i<oracleContracts.length; i++) {
             IOracle(oracleContracts[i]).init(interimTokens[i]);
         }
         emit OracleInitialized(msg.sender, baseToken, indexToken);
@@ -49,7 +49,7 @@ contract ICHICompositeOracle is OracleCommon {
      @dev chain length is constrained by gas
      */
     function update(address /* token */) external override {
-        for(uint i=0; i<oracleContracts.length; i++) {
+        for(uint256 i=0; i<oracleContracts.length; i++) {
             IOracle(oracleContracts[i]).update(interimTokens[i]);
         }
         // no event, for gas optimization
@@ -63,11 +63,11 @@ contract ICHICompositeOracle is OracleCommon {
      @param amountUsd index tokens required, precision 18
      @param volatility overall volatility metric - for future use-caeses
      */
-    function read(address /* token */ , uint amountTokens) public view override returns(uint amountUsd, uint volatility) {
-        uint compoundedVolatility;
-        uint amount = tokensToNormalized(interimTokens[0], amountTokens);
+    function read(address /* token */ , uint256 amountTokens) public view override returns(uint256 amountUsd, uint256 volatility) {
+        uint256 compoundedVolatility;
+        uint256 amount = tokensToNormalized(interimTokens[0], amountTokens);
         volatility = 1;
-        for(uint i=0; i<oracleContracts.length; i++) {
+        for(uint256 i=0; i<oracleContracts.length; i++) {
             ( amount, compoundedVolatility ) = IOracle(oracleContracts[i]).read(interimTokens[i], normalizedToTokens(interimTokens[i], amount));
             volatility = volatility.mul(compoundedVolatility);
         }
@@ -80,8 +80,8 @@ contract ICHICompositeOracle is OracleCommon {
      @param amountTokens tokens required in tokens native precision
      @param volatility metric for future use-cases
      */
-    function amountRequired(address /* token */, uint amountUsd) external view override returns(uint amountTokens, uint volatility) {
-        uint tokenToUsd;
+    function amountRequired(address /* token */, uint256 amountUsd) external view override returns(uint256 amountTokens, uint256 volatility) {
+        uint256 tokenToUsd;
         (tokenToUsd, volatility) = read(NULL_ADDRESS, normalizedToTokens(indexToken, PRECISION)); 
         amountTokens = PRECISION.mul(amountUsd).div(tokenToUsd);
         amountTokens = normalizedToTokens(indexToken, amountTokens);
@@ -95,7 +95,7 @@ contract ICHICompositeOracle is OracleCommon {
     /**
      @param count number of interim oracles
      */
-    function oracleCount() external view returns(uint count) {
+    function oracleCount() external view returns(uint256 count) {
         return oracleContracts.length;
     }
 
@@ -105,7 +105,7 @@ contract ICHICompositeOracle is OracleCommon {
      @param oracle interim token oracle address
      */
 
-    function oracleAtIndex(uint index) external view returns(address oracle, address token) {
+    function oracleAtIndex(uint256 index) external view returns(address oracle, address token) {
         return(oracleContracts[index], interimTokens[index]);
     }
 }
